requirements-dev.txt
-r requirements.txt
pytest
bump-pydantic==0.6 # (in dev as it is one time use)
